+++ CHANGELOG.md
@@ -8,8 +8,13 @@
 
 ### Node JS API
 
-* Add support for `lineFeed`, `indentWidth`, and `indentType` options to
+* Add support for `data`, `lineFeed`, `indentWidth`, and `indentType` options to
   `render()` and `renderSync()`.
+
+### Dart API
+
+* Add a `renderString()` method for rendering Sass source that's not in a file
+  on disk.
 
 ## 1.0.0-beta.1